requirements.txt
Django==5.1.6
djangorestframework==3.15.2
python-dotenv==1.0.1
drf-flex-fields==1.0.2
drf-extensions==0.7.1

celery
redis==5.2.1
django-redis==5.4.0

ippanel==2.0.7
django-import-export
django-currentuser==0.8.0
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
djangorestframework-simplejwt==5.4.0
psycopg2-binary
django-filter